+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - `--programs-list` argument
+- `--faces-list` argument
 ### Changed
 - `japanese-green-tea` program bug fixed
 - `README.md` updated